api/requirements.txt
flask
mysql-connector-python
python-dotenv
flask-sock

matplotlib
requests